requirements.txt
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
pipenv==2022.10.25
platformdirs==2.5.2
protobuf==3.20.1
PyMySQL==1.0.2
PyYAML==6.0
virtualenv==20.16.6
virtualenv-clone==0.5.7
Werkzeug==2.2.2
WTForms==3.0.1
